Update format string for send_after parameter (#71)
The UTC time offset is sufficient to specify a time.

`t` indicates the timezone itself. For example, in `Melbourne/Australia` where I live, `t` is parsed as `AEST`. The old string would be `2015-09-24 14:00:00 AEST+1000`, which is incorrect because AEST by itself is `UTC+1000`. This caused OneSignal to double-handle the time zone (turning `+1000` into `+2000`)